requirements.txt
Flask==3.1.0
ollama==0.4.7
pillow==11.1.0